Спринт 19. Заменил SettingsActivity на SettingsFragment.
