+++ app/src/main/res/layout/fragment_settings.xml
@@ -3,18 +3,15 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical">
+    android:orientation="vertical"
+    android:theme="@style/Theme.PlaylistMaker.Settings"
+    >
 
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginStart="@dimen/margin_20"
         android:orientation="horizontal">
-
-        <ImageView
-            android:contentDescription="@string/icon_back"
-            android:id="@+id/button_go_back"
-            style="@style/IconGoBack"/>
 
         <TextView
             android:text="@string/settings"
@@ -23,7 +20,7 @@
     </FrameLayout>
 
     <Switch
-        android:id="@+id/switch1"
+        android:id="@+id/buttonThemeSwitcher"
         android:layout_marginTop="@dimen/margin_61"
         android:text="@string/black_theme"
         style="@style/Switch" />
